chore: Split simple columns permutations for the Table
